+++ src/lib/header-bidding/utils.ts
@@ -40,6 +40,11 @@ const contains = (
 	size: HeaderBiddingSize,
 ): boolean => Boolean(sizes.find((s) => s[0] === size[0] && s[1] === size[1]));
 
+const isValidPageForMobileSticky = (): boolean => {
+	const { contentType, pageId } = window.guardian.config.page;
+	return contentType === 'Article' || pageId.startsWith('football/');
+};
+
 /**
  * Cleans an object for targetting. Removes empty strings and other falsy values.
  * @param o object with falsy values
@@ -197,7 +202,7 @@ export const shouldIncludeMobileSticky = once(
 			max: 'mobileLandscape',
 		}) &&
 			!isInUk() &&
-			window.guardian.config.page.contentType === 'Article' &&
+			isValidPageForMobileSticky() &&
 			!window.guardian.config.page.isHosted),
 );